requirements.txt
streamlit
langchain-nvidia-ai-endpoints
python-dotenv
PyPDF2
faiss-cpu
langchain_google_genai
langchain-text-splitters   
langchain-community
openai